complete Today weather card fix #75
