+++ src/route/index.js
@@ -7,6 +7,7 @@ import TodoLogin from '../components/TodoLogin'
 import TodoContent from '../components/TodoContent'
 import TodoMypage from '../components/TodoMypage'
 import TodoWeather from '../components/TodoWeather'
+import TodoSidebar from '../components/TodoSidebar'
 
 const router = new VueRouter({
     mode: 'history',
@@ -15,8 +16,8 @@ const router = new VueRouter({
         {path:'/todo', component: TodoContent},
         {path: '*', component: TodoLogin},
         {path: '/mypage', component: TodoMypage},
-        {path: '/weather', component: TodoWeather}
-
+        {path: '/weather', component: TodoWeather},
+        {path: '/Sidebar', component: TodoSidebar}
     ]
 })
